requirements.txt
fastapi
uvicorn
pytesseract
Pillow
transformers
sqlite3